+++ zerver/lib/event_schema.py
@@ -1786,7 +1786,7 @@ group_type = DictType(
         ("direct_subgroup_ids", ListType(int)),
         ("description", str),
         ("is_system_group", bool),
-        ("can_mention_group", int),
+        ("can_mention_groups", ListType(int)),
     ]
 )
 
@@ -1833,7 +1833,7 @@ user_group_data_type = DictType(
     optional_keys=[
         ("name", str),
         ("description", str),
-        ("can_mention_group", int),
+        ("can_mention_groups", ListType(int)),
     ],
 )